moved getOpenGraphData function to utils and switched from fetch to axios
